POST pour les nested params et factorisation user et currentuser api avec outdoor_controller api
